+++ Model/EF/ProductCategory.cs
@@ -9,7 +9,7 @@ namespace Model.EF
     [Table("ProductCategory")]
     public partial class ProductCategory
     {
-        public long ID { get; set; }
+        public int ID { get; set; }
 
         [StringLength(250)]
         public string Name { get; set; }
@@ -17,7 +17,7 @@ namespace Model.EF
         [StringLength(250)]
         public string MetaTitle { get; set; }
 
-        public long? ParentID { get; set; }
+        public int? ParentID { get; set; }
 
         public int? DisplayOrder { get; set; }
